prefixes: move `IdIndex` to `id_prefix` module
I'll reuse it there next.
